requirements.txt
python-telegram-bot
google-generativeai
requests
python-dotenv
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
